backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
primer3-py==1.0.0
biopython==1.81
requests==2.31.0
matplotlib==3.8.2
python-multipart==0.0.6
numpy==1.26.2
pydantic==2.5.2 